pyproject.toml
[tool.poetry]
name = "cal_app_back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.2"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pre-commit = "^4.0.1"
sqlalchemy = "^2.0.36"
poetry-dotenv = "^0.4.0"
python-dotenv = "^1.0.1"
dependency-injector = "^4.43.0"
alembic = "^1.14.0"
pymysql = "^1.1.1"
aiomysql = "^0.2.0"
pytest = "^8.3.4"
greenlet = "^3.1.1"
pytz = "^2024.2"
pyjwt = "^2.9.0"
pymongo = "^4.10.1"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
pytest-asyncio = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
